CollectionsUtility - SortedIntList - Intersect, Union
CollectionsUtility contains some collection classes and algorithms that are generally useful without specific ties to the rest of the project.

The two classes, SortedIntListIntersect and SortedIntListUnion, extends the LINQ Enumerable Intersect and Union operations to any number of lists.

Note that this commit merely contains a mock implementation and a smoke test that verifies the mock. A future commit will contain a highly optimized implementations specialized for sorted integer lists.